virt_underhill: rename to virt_mshv_vtl
Currently and somewhat fundamentally, `virt_underhill` depends on the
mshv_vtl driver, even for SNP and TDX. When we want to support SNP/TDX
without the Microsoft hypervisor in the future, we will need to decide
whether the mshv_vtl driver will be extended to support that, or whether
we'll want to take a different approach.

But either way, in the meantime, this is a more accurate name for the
crate.