flowey: only build with LTO on macOS
To restore build performance of flowey on non-macOS systems, add a trampoline
binary to build flowey with LTO only on macOS.

Ideally, we'd work around this in a less invasive way, but cargo does not have
support for target-specific profile options, target-specific environment
variables, target-specific aliases, or anything else I can think of that would
help here.

(We could also have a config.toml that gets passed to flowey invocations and
sets a custom rust flag specifically for macos. However, due to other cargo
limitations, that would require that all flowey invocations occur at the root
of the workspace, which is annoying.)