openhcl_boot: specify boot CPUs when sidecar is enabled
If sidecar is enabled, then pass `boot_cpus=a,b,c` to the Linux kernel,
which tells it to boot only the base VP of each sidecar node.

This will be ignored by the current included Linux kernel. A concurrent
patch to add support for the boot_cpus option is pending.