nvme_driver: add a sequence number to high cid bits
Due to its use of a slab for tracking pending commands, the NVMe driver
aggressively reuses cids. This sometimes makes debugging difficult,
especially when correlating logs across multiple sources. It also
increases the chance that a NVMe device or driver bug will cause us to
incorrectly complete an in-flight IO.

To resolve this, use the high bits of the CID as a sequence number.
There are six unused bits, which is enough to at least get some
variability.