openhcl_boot: query VP indexes for VTL0
The hypervisor fails VP index queries for VTLs that are not enabled for
the target VP. Since VTL2 is not enabled for the APs at boot, this
causes failures on multi-VP VMs.

Just query VTL0 instead. The hardware IDs will always match since we
don't support changing APIC IDs at runtime.

To catch this in the future, update the virt_whp hypervisor emulator to
match this behavior.