openhcl/loader: move snp specific pages to new vtl2 reserved region (#304)
Previously, we put some SNP specific pages into the same VTL2 config
region. #265 broke SNP, because this region contains pages that _cannot
be_ zeroed, like the VMSA, CPUID pages and secrets pages.

Move those to a new region, that is marked as persisted to both
kernelmode and usermode. It is unclear to me if the kernel requires that
the CPUID page and secrets page are persisted, but treat them like the
VMSA for now.

Fixes #295