guide: document the use of CVM kernel config fragments
L1 kernels fail to boot when the host hypervisor enforces isolation
features for confidential VMs. The reason of the failed boot is not
obvious and debugging tools may provide no hits (e.g., WinDBG does
not show any kernel boot logs).

The kernel fails to boot because the necessary isolation features
are not enabled in the default hcl-*.config files. The script
merge-cvm-config.sh selects this features.

Document this extra step.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>