RFC: add optional disk table to the VMGS file
The disk table lists the disks that the host is allowed to offer, along
with their encryption information (cipher + key). If the host offers a
disk that is not in the list, fail to boot.

This is designed to support unenlightened guests, and is opt in via
environment variable.

To use, prepopulate the VMGS file with the appropriate disk information,
e.g.:

```
vmgstool disk-table -f vmgs.vmgs add --d 00000001-a4a3-11ef-9347-43f645bdedf8 -c xts-aes-256 -k key.bin
```

And in your OpenHCL configuration, ensure OPENHCL_USE_DISK_TABLE=1 is set.

If using an encrypted disk, be sure to build openvmm-hcl with the
`disk_crypt` feature.