disk_backend: add new layered disk
Add a new disk type `LayeredDisk` that is composed of multiple layers,
each of which can contribute an arbitrary subset of the sectors of the
composed disk.

This is useful for modeling differencing VHDs, the RAM diff disk, and
the upcoming sqlite-backed disk. It can be used to providing
disk snapshots and read and write caches.

In this PR, only the RAM disk and the full disk layers are defined.