disk_*: handle out of range accesses
Currently, the SCSI and NVMe emulators validate that a read/write/unmap
is in range before calling down to the backing disks. This creates a
TOCTOU problem because the disk size can change at any time,
asynchronously from any IO. It also creates additional synchronization
headaches since the disk implementations must be able to provide the
up-to-date disk size efficiently.

As a first step, add this range checking in the various disk
implementations. A subsequent change can remove the redundant checks in
the NVMe and SCSI emulators.