vmbus_client: use existing mesh channels to send revoke
In the vmbus client, instead of sending a revoke over the general notify
channel (which is hard to multiplex between multiple users of the vmbus
client), notify a user of a revoke by closing the per-channel
notification channel.

This is the last non-offer use of the notify channel. Change the notify
channel to just an offer channel.