mesh_channel: hide RPC response receiver from callers
In the future, RPC channels will typically (always?) use blocking
senders to provide backpressure and better support early cancellation.

To prepare for this, wrap the RPC response oneshot channel in a separate
type which in the future can contain both a blocking send request and
the response receiver.

Also, hide the internal fields of `Rpc` to give us flexibility there in
the future, too.