mesh: remove upcast/downcast concept
Mesh has a clever upcast/downcast type machinery to track whether a
given type has a compatible encoding with another type, and to allow you
perform conversions between types if they are compatible. This is pretty
nice--it allows you to erase and reclaim type safety for mesh messages
and especially channels.

However, this scheme adds a lot of complexity to the code. It falls over
in complicated cases where you'd like to use it. It hides some expensive
conversions that probably should be explicit. And in practice, it only
gets used in a few cases: the mesh worker infrastructure, mesh RPC
(gRPC/ttrpc), and the OpenHCL diag service.

The mesh cases are just about type erasure. We can do the same thing
manually without type safety and rely on tests to catch bugs; there's
nothing particularly tricky going on that warrants extra type checking.

The diag service case is because we want to swap a prost-generated type
for a mesh-generated type. Instead of doing this via upcast, update the
mesh service generator to allow replacing types at code generation time.
This is much cleaner anyway.