openhcl/page_pool_alloc: support save restore (#418)
Support save restore for the page_pool_alloc crate. This allows save
restore of the shared visibility pool.

A future PR will add support for a private page pool to be used by AK
cert renewal and other usages.

NOTE: This is a backport of #418 that does not include the underhill_core
changes. This is to allow future cherry-picks to apply cleanly, but we
do not want to actually support save restore of the pool in underhill_core,
as it's unneeded in release/2411.