openhcl: cleanup private page pool changes  (#461)
Some additional cleanup of the private pool and fixes with keep alive on
openvmm, and enable using the private pool via cmdline.

Do not save the pool state if we are not doing keep alive, as no device
should have allocations that need persisting. Fix advertising keep alive
support on openvmm because it doesn't work yet.

NOTE: Backport commit from main that has dropped changes that do not
matter for release/2411.