nvme,mana: ensure completions are read and written in order (#506)
The NVMe completion entry is a 16-byte value, and it contains a phase
bit that the NVMe driver reads to determine when a completion in the
queue is valid. To ensure the driver observes a consistent completion
entry, the phase bit must be written by the device _after_ the rest of
the entry, and it must be read by the driver _before_ the rest.

The memory access machinery in OpenVMM does not provide many ordering
guarantees, especially on ARM64. Ensure accesses are performed in the
right order via two atomic 8-byte accesses, plus proper fences, in the
NVMe emulator and driver.

There's a similar bug in MANA. Fix this, too.