pal_async: support running under miri
The epoll backend for `pal_async` registers an eventfd for waking up due to a waker, but it neglects to pass the `EPOLLET` (edge triggered) flag when registering it. Miri, it turns out, supports epoll, but it only supports `EPOLLET` registrations.

Pass `EPOLLET` for the wake event. This works fine--we are already handling the wake event in an edge-triggered compatible way.

This allows `#[async_test]` tests to run under miri on Linux.