guid: define macro for compile-time GUIDs
Instead of requiring users to be careful to use `from_static_str` only in const
context, offer a macro `guid!` that ensures that the guid is parsed in
at compile time.

This fixes some unnecessary runtime parsing of GUIDs in a few places,
and it gives us flexiblity to fork the compile-time and runtime parsing
implementations if we want to.