petri: trace the test result
Right now, a test result on failure (the panic text or the anyhow
error) only gets written to stdout, not to the split text logs or to the
JSON log. This makes it hard to diagnose test failures by just looking
at the log files.

Trace the test result to the other logs, too. Also, add a file to make
it easy to find the tests that passed and failed in the test result
artifact directory tree without having to parse the junit XML.