vmbus_server: don't send synic messages while paused
The synic may be saved or reset while the vmbus server is paused. Don't
send any messages during this time, since this can result in torn or
corrupt state.

Skip this for vmbus running inside OpenHCL for now, since this relies on
the newer saved state field for messages queued by vmbus, and
release/2411 won't process that field.