vm_topology: remove physical address width from MemoryLayout
The physical address width is not really a singular thing--there are
many different interesting address widths:

* The width the virtual processor can access via CPU instructions.
* The width virtual/mapped devices/the IOMMU can access via DMA.
* The width the hypervisor supports for mapping memory (usually this is
  limited to the min of the previous two)
* The width reported by cpuid on x86.
* The width reported by sysregs on arm64 (this will be rounded up to the
  next multiple of 4).
* The offset of the VTL0 alias map.
* The size of the VA range we allocate for guest memory mappings.

Some of these have ended up getting conflated in our designs since these
are often the same/related to each other in useful ways. But in
practice, there are subtle differences. And in reality, we don't really
need the value in as many places as we think.

So, remove the generic, underspecified physical address width from
`MemoryLayout`. Add a value to the x86 partition capabilities for use in
places where we need to specifically get the value that's reported to
the guest (for computing MTRRs). Remove propagation of the address width
from openhcl_boot through device tree.