tmk: add simple VMM for running TMKs
To make it easier to test core `virt_*` functionality without all the
accoutrements of a full-featured VMM (virtual devices, etc.), create a
VMM that can just run simple test microkernels (TMKs) and not full guest
operating systems.

This VMM can run either on the host or in a paravisor environment.

Current limitations:

* No docs on using this
* No integration with CI test pass
* x86_64 only
* Non-CVM only
* Just define one simple TMK that doesn't test much of anything

Ideally, this will mesh up with the effort to define some rich TMKs for
use with full-features VMMs, so that we can run some subset of TMKs in
both environments. But even without that, this will be valuable for
platform bringup.