flowey: merge minor rust steps in generated YAML
It is hard to read the output logs of a flowey-based pipeline since
there are so many steps (many of them implicit) that are just shuffling
variables around.

Furthermore, node authors avoid using variable combinators such as `map`
and `zip` since they generate so much noise in the pipeline output.

ADO pipelines partially sidestep these problems by merging all
contiguous Rust steps together, but GitHub pipelines don't do this. And
for ADO, it is difficult to debug the output because so many unrelated
but failure-prone steps get merged into a single ADO step.

To resolve these problems, add a new Rust step property: whether the
step can be merged with adjacent steps. Merge mergable steps with their
neighbors (whether the neighbor itself is mergable or not). Also merge
the generated steps for dealing with conditions, and eliminate the need
for putting conditions in the environment just to conditionalize another
flowey Rust step.

Mark most of the failure-free variable-shuffling steps as mergable
(termed "minor").

This greatly reduces the number of steps in the GH pipelines, but it
increases the number in the ADO pipelines. I think this is probably
worth the tradeoff.