virt_mshv_vtl/tdx: log more info for bad guest state exits (#1092)
VP.ENTER can fail if the guest's VMCS is in a bad state. This is a VMM
bug, so handle it rather specifically than just lumping it into the
generic exit type. Add some additional tracing to capture VMCS fields to
help debug issues.

Fix the VMX exit definition to correctly have the upper bits that show
different error conditions so we can correctly check if VP.ENTER
succeeded.

This is to help track down a current issue that has a very low repro
rate where we're seeing this exit code. There's an identified issue with
CR4 MCE (#1091), but it's unclear if that is the cause or not yet.