hyperv_ic: add KVP implementation
Add support for the KVP integration service. This can be used to query
and set "key-value pair" data in the guest OS and to query and set guest
networking configuration (IP addresses and the like).

Some versions of Linux, configured for Hyper-V or Azure, depend on this
being present even if the host does not use it.

Also, add a petri test for it, and add openvmm interactive console
commands for manual use of it.