vmbus_channel: Stopped vmbus devices should not see vmbus requests. (#1128)
During stop, the vmbus channels are stopped before the vmbus server,
which leaves a small window where a vmbus request from the guest can
arrive after the channel is stopped. Handle this by pending the message
in the vmbus channel code. These pended messages do not need to be
saved, because vmbus_server will replay them on restore. The maximum
number of pended requests should be two per channel_idx, for the case of
a pended modify request followed by a pended close. Any other
combination should be rejected by vmbus as the device will not be in a
valid state.

---------

Co-authored-by: Brian Perkins <brian.perkins@microsoft.com>