tmk: add a few APIC tests
Add a few APIC tests. So support this, add the ability to enable
exception and interrupt handlers on x86_64. Refactor the tmk core to a
separate crate for better separation between tests and test framework.

One of the new tests fails with `virt_support_apic`. I'll fix this in a
separate change so that it's easier to see. (These tests don't run in CI
yet so there won't be any breaks.)