flowey_lib_hvlite: ensure ordering of nextest publishing steps
I was performing some other refactoring and noticed some incorrect
reordering in the generated pipeline. This was due to the variables
containing the output paths for VMM tests not depending on the VMM tests
actually having run.

Add in this extra dependency, using a new method on `ReadVar` to make
this easy and relatively clear.