gen_init_ramfs: set all modification times to 0
The generated openhcl initrd file has embedded modification times.
Remove these so that subsequent invocations of build-igvm produce
an identical binary.

A less aggressive change would have set the mtime to the time of the
config file or something. But these modification times really have no
value for us, and in the future the config file itself may be generated
at runtime by flowey (if it's not already--I didn't check). So just use
zero (Unix epoch).