flowey: make values secret, not variables
Currently, a `ReadVar`/`WriteVar` pair can be marked as secret, in which
case flowey is careful never to display it in logs. To mark a variable,
the user must remember to create the variable pair with `new_secret_var()`,
and the user must ensure that users of the variable to do not rewrite
its contents into some other, non-secret variable. This is hard to do
accurately, especially as we change the code to create more variables
implicitly.

Change the model so that _variables_ are not secret but their _values_
can be--when any variable is written to, the caller can specify that the
value is secret. Propagate this to readers of the variable, even if this
variable is converted into and back from a CI environment variable.

By default, be conservative: once a Rust step reads a secret from a
variable, treat all future values written as secret. Add specific
`write_secret` and `write_non_secret` methods for overriding this
default.