inspect: reduce object sizes
Wrap `Deferred` in a `Box` so that it's cheaper to pass around. Move
immutable parts of `Request` into a structure that can be stored at
the base of the call stack. Remove some deferred initialization
in `Response`.

As part of this, eliminate the `Response::request` method for getting
a new `Request` from an existing `Response`. Instead, use the `merge`
method along with `adhoc`. This is necessary to avoid needing extra
checks in the common path. It's also more consistent with other uses of
response.