cvm_tracing: Allow untagged tracing statements by default (#1402)
After long discussions we have decided to flip the default of our
tracing filter, and to allow untagged tracing statements by default. We
believe that the risks and costs of being unable to debug incidents in
production are too high, and that we can manually scrub our tracing
statements to ensure that no sensitive information is leaked.

Part of #852.