sidecar: zero initialize unmapped register page (#1440)
The register page is not valid until it has been mapped and the
hypervisor says it's valid. However, the in-memory contents of the page
before it is mapped could be non-zero, especially across a servicing
operation. This can cause the VMM to think the register page contents
are valid, causing register corruption.

Fix this by explicitly clearing the valid flag in the register page
during sidecar startup.

Also zero a few more potentially stale structures to avoid potential
bugs.