underhill_core: serialize sidecar VP online
The Linux kernel serializes CPU online. If multiple sidecar VPs need
to be onlined into OpenVMM simultaneously, they will all stop running
the guest OS and wait for their turn to be onlined into Linux.
This means the average blackout time for a VP that's onlined early in
boot is linear in the number of early-onlined VPs. And thanks to
typical device configurations, this is usually linear in the total number of VPs.

To avoid this, explicitly serialize VP online _before_ the target VP
is stopped. This allows the VP to continue running the guest until it
reaches the front of the queue. This reduces the average blackout time to
just the time to online one CPU, meaning this solution should scale
to any number of VPs.