nvme: support multiple submission queues per completion queue
The NVMe spec requires that multiple submission queues be able to target the
same completion queue. Initially we thought no guests really rely on this, but
it turns out that depending on the VM's topology, Windows expects this to work.

Support this without adding extra synchronization--have a single task own all
submission queues that are tied to a single completion queue. This limits CPU
parallelism, but I don't really expect that to be the bottleneck for common use
cases, since this configuration implies the guest has a single CPU handling
completions anyway.