sparse_mmap: fail at compile time on unaligned try_cmpxchg inputs
Prevent UB by requiring the input type to `try_compare_exchange` to
be naturally aligned. Remove the error-prone `try_compare_exchange_ref`
function and update callers to explicitly match on the size of their
input.

With some additional work, we could instead choose to avoid requiring
alignment for these input types (since they're only accessed via pointer
as part of the calling convention with the assembly routines to perform
the cmpxchg). But this is not a significant constraint on callers, so
take the simpler route for now.