nvme_driver: don't flr nvme devices (#1714)
The default `vfio` device behavior is to issue a function level reset
when attaching or detaching devices. It does so because the device is in
an unknown or untrusted state. However, within the context of a trusted
virtualization stack, OpenHCL can reasonably trust the state and
behavior of the device. So, optimize performance by removing these
function level resets for nvme devices. This follows the same model as
already exists for MANA devices.

The `nvme_driver` already shuts down the device (see
`NvmeDriver::reset()`) and waits for the device to become disabled. A
well behaved nvme device will not issue DMA after this point. That same
device should tolerate a graceful start without an FLR.

Pending work before this PR is ready to commit:

- [x] Initial poc
- [x] Parameterize via command line (provide a way to disable, and also
easily run A/B tests)
- [x] Fix
https://github.com/microsoft/openvmm/pull/1714#issuecomment-3085846035
- [x] Check no regressions on CI
- [x] Test servicing locally with real NVMe devices