virt_mshv_vtl/mshv: Limit crash message size (#1953)
When guests are in the middle of crashing they can be in broken states
and do weird things. We've seen cases where they pass some outrageously
large number to this (or possibly a negative number?), which results in
us attempting to alloc a giant Vec, which results in us OOMing. Limit
this to a single page to match the spec, and the hypervisor.