vmbus_server: improve hvsock connection feedback
In some cases, Linux guests fail to open hvsock connect offers. This
manifests as a hang with the host-side Unix socket connection open, plus
an offer leak.

Fix this by adding a timeout waiting for the guest to connect and by
only connecting to the Unix socket once the vmbus channel open process
is further along.

This should improve pipette connection reliability for petri tests.