netvsp: batch calls to tx_avail
Instead of calling `tx_avail` for each vmbus packet message, accumulate
all the packets that are currently in the ring buffer and transmit them
at once. This reduces the number of doorbell writes to the mana device.

Also, clean up the hot path a bit to avoid a bunch of extra checks.
Split the packet path out from the control message path.