minircu: improve codegen
Restructure some code so that rust/LLVM can do a better job optimizing
it.

The most important change seemed to be to avoid capturing the
return value of `ThreadData::run()` before exiting the RCU critical
section and instead using a `Drop` impl to perform the exit. This is
better anyway, if `minircu` is ever used with unwinding panics.

This change improves some guest memory access benchmarks by 30%.