trycopy: reimplement sparse_mmap's trycopy with inline asm
Use inline asm instead of global asm or a C implementation, which
allows the compiler to inline the code. Use a
custom exception-table-based approach for recovery instead of
setjmp/longjmp on Unix (slow) or SEH exceptions on Windows (prevents
inlining).

Besides improving performance across the board, this allows us to use
a single per-arch implementation of the primitives across all OSes.
This should make it easier to add new primitives over time, if desired.