pal_async: allow dropping overlapped files with in-flight IO
When a `IoOverlapped` is dropped, we disassociate the underlying file
from its completion port so that the file can be reused for other
purposes. Windows does not support this while there is in-flight IO,
so if the drop happens before all IO has completed, we panic.

However, in most use cases, we don't need to explicitly disassociate,
since the `File` is going to be dropped anyway, and it _is_ generally
OK to close a file handle with IO in flight. Change things around so
that we only disassociate when the caller explicitly wants to reuse
the file, and make the contract clear that this will panic if there
is in-flight IO.

Also, make the `OverlappedFile` safety contract clearer (if not
particularly easy to ensure--Windows does not make this easy).