openhcl_boot: nvme keepalive flag passing cleanup (#2445)
The disable keepalive flag seems to be set in several places right now
and can be a little bit confusing to follow along with.
`BuildKernelCommandLineParams` seems to be containing both the
`PartitionInfo` and the `nvme_keepalive` flag from command line options
(which is derived from `PartitionInfo`). This can be streamlined.