CI: Fully fix the unsafe check by going back to pull_request_target and rewriting (#2507)
Was hoping this wouldn't be necessary, as it allows more permissions
than we'd like, but it seems to be the only way to really have write
access to make comments and modify labels. Rewrite the script to fetch
content via github's API, instead of checking out to the filesystem, to
avoid any filesystem-based attacks.