underhill_core: specify which CPUs have outstanding IO, not just if they have interrupts (#2512)
This PR supports upcoming changes to to have sidecar start CPUs without
outstanding IO, but have the kernel start the CPUs with outstanding IO.
A 96-vp VM will possibly have 96 CPUs with interrupts, but only have IO
outstanding to a few of those at a time. This change allows the restore
path to discern.

TEST:
Validated that this does not cause regression in local CI.