nvme_resources: add a mechanism to verify enable/disable behavior for PCI events (#2545)
Adds an "Enable" verification mechanism on the PCI side similar to we
verify nvme commands; by using `OneshotSender` to communicate with the
test that a certain command or event was seen.
A few other changes along as support changes:
* CancelContext requires a double unwrap which some tests were not
doing.
* Fixed some nit comments from previous PRs 
* `apply_keepalive_fault` function now returns the created vm object.
This fixes the issue that I was observing where the sender was dropped
after servicing.