disk_striped: optimize futures vectors (#2429)
Performance traces show that growing and populating the vector of
futures in `disk_striped` IO functions is expensive. Optimize this: use
`Iterator::collect` to collect directly into the `JoinAll` that we'll
await, using an iterator that implements `TrustedLen` so that the vector
can allocate its buffer once and construct the futures in place.

Also, clean up some impossible failure paths.

---------

Co-authored-by: Matt LaFayette (Kurjanowicz) <mattkur@microsoft.com>